refactor: replace league typescript types by using actual foundry code
This commit introduces a shell script that copies all foundry code from a docker image into the project.
The code will then be interpreted by typescript, reading the jsdoc comments in it.
This works for some stuff, but not everything. For some cases I opted to override it with leniant
definitions that use any instead of proper typing. At least for now. This might change with future updates.