fix: hide informations from players that should not be visible to them
This was a big one. There was a problem when it came to handling private information
when a related entity was private but not the connection to it. For this I refactored
the whole page creation to allow simpler and improved handling of ownership as well as
resolving and counting private and public relations.

Fixes #136