feat!: Update entity browser to use ApplicationV2 framework from Foundry
BREAKING CHANGE: This raises the minimum foundry version from v11 to v12. The whole entity browser
was basically recreated to work with the new ApplicationV2 framework. This changes
the styles and tries to keep close to base foundry styles. It should work with
light and dark themes - and hopefully others too.
One feature thatw as removed is the overview of linked but since in Kanka deleted
journal entries. You'll have too keep track of those yourselfes from now on.

Fixes #149